outsource PackageCell code to seperate file
